chore: organize vector store folder and files
