--- src/services/vector_store/__init__.py
@@ -1,12 +1,0 @@
-# src/services/vector_store/__init__.py
-"""Vector store module for Alfred Bot"""
-
-from .base import VectorStore
-from .chroma_store import ChromaVectorStore
-from .search_service import VectorSearchService
-
-__all__ = [
-    "VectorStore",
-    "ChromaVectorStore",
-    "VectorSearchService",
-]
